Moved RegTest-specific code into its own file
